List of all functions

Here an overview of all functions:

GeoModBox.Geometry — Type

Geometry()

Structure to initialize the geometry of the model domain.

GeoModBox.HeatEquation.OneD.AssembleMatrix1Dc! — Method

AssembleMatrix1Dc!( κ, Δx, Δt, nc, BC, K )

Setup the coefficient matrix for the linear system of equations.

GeoModBox.HeatEquation.OneD.BackwardEuler1Dc! — Method

BackwardEuler1Dc!( implicit, κ, Δx, Δt, nc, BC , K)

Solves the onedimensional heat diffusion equation assuming no internal heating and constant thermal parameters using an implicit, backward euler finite difference scheme.

The temperature is defined on central nodes and the heat flux on the vertices. Boundary conditions are currently limited to Dirichlet and Neumann. Using central temperature nodes requires external ghost nodes, which are used to define the boundary conditions.

implicit    : Tuple, containing the current temperature array T0 and 
              the new temperature array T
κ           : Diffusivity [ m²/s ]
Δt          : Time step [ s ]
nc          : Number of central nodes
Δx          : Grid spacing [ m ]
BC          : Tuple for the boundary condition
K           : Coefficient matrix for linear system of equations

GeoModBox.HeatEquation.OneD.CNA1Dc! — Method

CNA1Dc!( cna, κ, Δx, Δt, nc, BC, K1, K2 )

Solves the onedimensional heat diffusion equation assuming no internal heating and constant thermal parameters using Crank-Nicolson finite difference scheme.

The temperature is defined on central nodes and the heat flux on the vertices. Boundary conditions are currently limited to Dirichlet and Neumann. Using central temperature nodes requires external ghost nodes, which are used to define the boundary conditions.

cna         : Tuple, containing the current temperature array T0 and 
              the new temperature array T
κ           : Diffusivity [ m²/s ]
Δt          : Time step [ s ]
nc          : Number of central nodes
Δx          : Grid spacing [ m ]
BC          : Tuple for the boundary condition
K1          : Coefficient matrix for the unknow variables 
K2          : Coefficient matrix for the know variables

GeoModBox.HeatEquation.OneD.ComputeResiduals1Dc! — Method

ComputeResiduals1Dc!( cna, κ, Δx, Δt, nc, BC, K1, K2 )

Computes the residual of the onedimensional heat diffusion equation assuming no internal heating and constant thermal parameters.

The temperature is defined on central nodes and the heat flux on the vertices. Boundary conditions are currently limited to Dirichlet and Neumann. Using central temperature nodes requires external ghost nodes, which are used to define the boundary conditions.

dc          : Tuple, containing the current temperature array T, 
              the temperature array with ghost nodes T_ex,
              the partial derivatives ∂T2∂x2, and the
              residual R
κ           : Diffusivity [ m²/s ]
Δx          : Grid spacing [ m ]
Δt          : Time step [ s ]       
BC          : Tuple for the boundary condition

GeoModBox.HeatEquation.OneD.ForwardEuler1D! — Method

ForwardEuler1D!( explicit, κ, Δx, Δt, nc, BC)

Solves the onedimensional heat diffusion equation assuming internal heating and variable thermal parameters using an explicit, forward euler finite difference scheme.

The temperature is defined on central nodes and the heat flux on the vertices. Boundary conditions are currently limited to Dirichlet and Neumann. Using central temperature nodes requires external ghost nodes, which are used to define the boundary conditions.

T           : Tuple, containing the regular temperature array T and 
              array containing the ghost nodes T_ex
Py          : Tuple, containing the thermal parameters ρ, k, cp, and H [ W/kg ]
Δt          : Time step [ s ]
Δy          : Grid spacing [ m ]
nc          : Number of central nodes
BC          : Tuple for the boundary condition

GeoModBox.HeatEquation.OneD.ForwardEuler1Dc! — Method

ForwardEuler1Dc!( explicit, κ, Δx, Δt, nc, BC)

Solves the onedimensional heat diffusion equation assuming no internal heating and constant thermal parameters using an explicit, forward euler finite difference scheme.

The temperature is defined on central nodes and the heat flux on the vertices. Boundary conditions are currently limited to Dirichlet and Neumann. Using central temperature nodes requires external ghost nodes, which are used to define the boundary conditions.

explicit    : Tuple, containing the regular temperature array T and 
              array containing the ghost nodes T_ex
κ           : Diffusivity [ m²/s ]
Δt          : Time step [ s ]
nc          : Number of central nodes
Δx          : Grid spacing [ m ]
BC          : Tuple for the boundary condition